cache request body to avoid attempting to read from exhausted stream
The request body is read from the request stream and it can only be
consumed once. FastAPI's Request class does cache the body so it can be
read multiple times. The problem, however is that the request instance
that the middleware uses is not the same that will be passed to the view
consequently if we try to read the request body in the middleware and
the view has already read the body as well, the stream will have already
been consumed and the application will hang. To fix this, we wrap the
receive function which reads from the stream so that it caches the body
if it is read in the view.